take GetKeychain back to deploycmd
