+++ cmd/subnetcmd/deploy.go
@@ -20,9 +20,11 @@ import (
 	"github.com/ava-labs/avalanche-cli/pkg/subnet"
 	"github.com/ava-labs/avalanche-cli/pkg/txutils"
 	"github.com/ava-labs/avalanche-cli/pkg/ux"
+	ledger "github.com/ava-labs/avalanche-ledger-go"
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/crypto/keychain"
 	"github.com/ava-labs/avalanchego/utils/formatting/address"
+	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/vms/platformvm/txs"
 	"github.com/ava-labs/coreth/core"
 	spacesvmchain "github.com/ava-labs/spacesvm/chain"
@@ -260,7 +262,7 @@ func deploySubnet(cmd *cobra.Command, args []string) error {
 	// from here on we are assuming a public deploy
 
 	// get keychain accesor
-	kc, err := key.GetKeychain(useLedger, app.GetKeyPath(keyName), network)
+	kc, err := GetKeychain(useLedger, app.GetKeyPath(keyName), network)
 	if err != nil {
 		return err
 	}
@@ -598,4 +600,45 @@ func PrintReadyToSignMsg(
 	ux.Logger.PrintToUser("")
 	ux.Logger.PrintToUser("Commit command:")
 	ux.Logger.PrintToUser("  avalanche transaction commit %s --input-tx-filepath %s", chain, outputTxPath)
+}
+
+func GetKeychain(
+	useLedger bool,
+	keyPath string,
+	network models.Network,
+) (keychain.Keychain, error) {
+	// get keychain accesor
+	var kc keychain.Keychain
+	if useLedger {
+		ledgerDevice, err := ledger.New()
+		if err != nil {
+			return kc, err
+		}
+		// ask for addresses here to print user msg for ledger interaction
+		ux.Logger.PrintToUser("*** Please provide extended public key on the ledger device ***")
+		addresses, err := ledgerDevice.Addresses(1)
+		if err != nil {
+			return kc, err
+		}
+		addr := addresses[0]
+		networkID, err := network.NetworkID()
+		if err != nil {
+			return kc, err
+		}
+		addrStr, err := address.Format("P", key.GetHRP(networkID), addr[:])
+		if err != nil {
+			return kc, err
+		}
+		ux.Logger.PrintToUser(logging.Yellow.Wrap(fmt.Sprintf("Ledger address: %s", addrStr)))
+		return keychain.NewLedgerKeychain(ledgerDevice, numLedgerAddressesToDerive)
+	}
+	networkID, err := network.NetworkID()
+	if err != nil {
+		return kc, err
+	}
+	sf, err := key.LoadSoft(networkID, keyPath)
+	if err != nil {
+		return kc, err
+	}
+	return sf.KeyChain(), nil
 }
